pyproject.toml
[tool.poetry]
name = "lucrezia-ai-news"
version = "0.1.0"
description = ""
authors = ["panosAthDBX <panos.athanasiou+github@databricks.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
bs4 = "^0.0.2"
transformers = "^4.42.4"
google-api-python-client = "^2.137.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
